requirements.txt
PyYAML
SPARQLWrapper
SQLAlchemy
GeoAlchemy2
aiohttp
aiohttp_utils
astral
psycopg2
pytz
rdflib
shapely